Memorize a type through which a method was accessed
If the type of an object on which a method was called doesn't have an
appropriate definition, thus initiating inheritance lookup, then we now
pass the initial type along the way until one of the two things happen:

1. Some base type of the initial type has a definition that match. Then
just forget the initial type.

2. No parents of the initial type have an appropriate definition. Then
type object will return $Failed and print a message containing the type
which was accessed rather than a real type of an object. It makes the
message more informative to users because what real type implements
the type through which a user accessed the method is an implementation
detail.